// src/parity/utils/types.ts
export interface EventCommand {
  code: number;
  indent: number;
  parameters: unknown[];
}